Allow apps to use different configs (#143)
* Create generic webpack config

* Add webpack config for gp-v1

* Create config for v2 apps

* Use footer for v2

* Remove envs for webpack config

* Add envs to default webapack config

* Add envs for gp-v1

* Use special webpack config for gp-v1

* Improve assert message


* Remove dead code
