+++ package.json
@@ -11,14 +11,14 @@
     "start": "npm run start:explorer",
     "start:explorer": "APP=explorer npm run start:apps",
     "start:trade": "APP=trade npm run start:apps",
-    "start:gp-v1": "APP=gp-v1 npm run start:apps",
+    "start:gp-v1": "APP=gp-v1 npm run start:apps -- --config webpack.gp-v1.config.js",
     "start:apps": "NODE_ENV=development webpack-dev-server --mode development --hot --host 0.0.0.0 --disable-host-check",
     "start:dist": "serve dist -s -l 8081",
     "mock": "MOCK=true yarn start",
     "build": "rimraf ./dist && NODE_ENV=production webpack --mode production",
     "build:explorer": "APP=explorer npm run build",
     "build:trade": "APP=trade npm run build",
-    "build:gp-v1": "APP=gp-v1 npm run build",
+    "build:gp-v1": "APP=gp-v1 npm run build -- --config webpack.gp-v1.config.js",
     "lint:js": "eslint src --cache --cache-location .eslint-cache --ext .ts,.tsx,.js",
     "lint:js:fix": "eslint src --fix --ext .ts,.tsx,.js",
     "lint:css": "stylelint --cache --cache-location .stylelint-cache 'src/**/*.[jt]s?(x)' --syntax 'css-in-js'",
