+++ src/components/layout/GenericLayout/Footer/config.ts
@@ -2,8 +2,10 @@ export const footerConfig = {
   verifiedText: 'View verified contract',
   isBeta: true,
   url: {
-    web: 'https://github.com/gnosis/dex-react/tree/v',
-    appId: 'https://github.com/gnosis/dex-react/wiki/App-Ids-for-Forks',
-    contracts: 'https://github.com/gnosis/dex-contracts/tree/v',
+    web: 'https://github.com/gnosis/gp-ui/tree/v',
+    // TODO: Pending to move and adapt the wiki
+    appId: null,
+    // appId: 'https://github.com/gnosis/gp-v1-ui/wiki/App-Ids-for-Forks',
+    contracts: `https://github.com/gnosis/gp-v2-contracts/tree/v`,
   },
 }
