+++ src/components/layout/GenericLayout/Footer/index.tsx
@@ -104,8 +104,8 @@ export const Footer: React.FC<FooterType> = (props) => {
             Web: v{VERSION}
           </a>
         ) : null}
-        {url.appId && CONFIG.appId ? (
-          <a target="_blank" rel="noopener noreferrer" href={url.appId}>
+        {CONFIG.appId ? (
+          <a target="_blank" rel="noopener noreferrer" href={url.appId ?? '#'}>
             App Id: {CONFIG.appId}
           </a>
         ) : null}
